requirements.txt
streamlit==1.24.1
openai==0.27.8
SQLAlchemy==2.0.18
faker==18.9.0
python-dotenv==1.0.0
pydantic==2.3.0